@@ -10,7 +10,7 @@ authors = [
     {name = "k-ush", email = "argoooooon@gmail.com"},
 ]
 readme = "README.md"
-requires-python = ">=3.8,<4.0"
+requires-python = ">=3.9,<4.0"
 classifiers = ["Typing :: Typed"]
 urls = {homepage = "https://github.com/argonism/mdfy", repository = "https://github.com/argonism/mdfy"}
 dependencies = [
